Fix homework-2

+++ assignments/24-HR-CS1/homework 2/Koptev/Homework2.py
@@ -1,5 +1,6 @@
-startNum = float(input("Enter a start number: "))
-interestRate = float(input("Enter an interest rate (%): ")) / 100
-period = float(input("Enter a period (year): "))
+from decimal import Decimal
+startNum = Decimal(input("Enter a start number: "))
+interestRate = Decimal(input("Enter an interest rate (%): ")) / 100
+period = Decimal(input("Enter a period (year): "))
 income = startNum * interestRate * period
 print("The income is:", income)
